+++ closetr-api/api/routes/clothes/index.js
